Extract captcha logic into new module
In preparation for the dynamic captcha feature, the captcha creation
logic is extracted into it's own module. This makes it easier to reuse
the specific parts needed.
